refactor(ChessBoard): KingFinder의 메서드를 ChessBoard로 옮김
